Move update_lightbox_caption Turbo Stream action to Stimulus controller
Refactored the custom update_lightbox_caption Turbo Stream action from
application.js into a dedicated Stimulus controller to follow Stimulus
organizing principles where element-specific JavaScript should be
scoped to a controller bound to an element.

Changes:
- Created lightbox_caption_sync_controller.js that registers the custom
  Turbo Stream action on connect and cleans it up on disconnect
- Removed Turbo.StreamActions.update_lightbox_caption from application.js
- Added lightbox-caption-sync controller to #content element in layout

This ensures the custom action is only registered when the controller
is connected to an element on the page, not globally in application.js.
